move solvers from Lab1 to parent dir and rename imports
