+++ solvers_lab1.py
